release v1.26.0

+++ src/main/java/com/cobo/waas2/model/SwapActivity.java
@@ -116,11 +116,11 @@ public class SwapActivity {
 
   public static final String SERIALIZED_NAME_CREATED_TIMESTAMP = "created_timestamp";
   @SerializedName(SERIALIZED_NAME_CREATED_TIMESTAMP)
-  private Integer createdTimestamp;
+  private Long createdTimestamp;
 
   public static final String SERIALIZED_NAME_UPDATED_TIMESTAMP = "updated_timestamp";
   @SerializedName(SERIALIZED_NAME_UPDATED_TIMESTAMP)
-  private Integer updatedTimestamp;
+  private Long updatedTimestamp;
 
   public static final String SERIALIZED_NAME_NETWORK_FEE = "network_fee";
   @SerializedName(SERIALIZED_NAME_NETWORK_FEE)
@@ -399,7 +399,7 @@ public class SwapActivity {
   }
 
 
-  public SwapActivity createdTimestamp(Integer createdTimestamp) {
+  public SwapActivity createdTimestamp(Long createdTimestamp) {
     this.createdTimestamp = createdTimestamp;
     return this;
   }
@@ -409,16 +409,16 @@ public class SwapActivity {
    * @return createdTimestamp
   **/
   @javax.annotation.Nullable
-  public Integer getCreatedTimestamp() {
+  public Long getCreatedTimestamp() {
     return createdTimestamp;
   }
 
-  public void setCreatedTimestamp(Integer createdTimestamp) {
+  public void setCreatedTimestamp(Long createdTimestamp) {
     this.createdTimestamp = createdTimestamp;
   }
 
 
-  public SwapActivity updatedTimestamp(Integer updatedTimestamp) {
+  public SwapActivity updatedTimestamp(Long updatedTimestamp) {
     this.updatedTimestamp = updatedTimestamp;
     return this;
   }
@@ -428,11 +428,11 @@ public class SwapActivity {
    * @return updatedTimestamp
   **/
   @javax.annotation.Nullable
-  public Integer getUpdatedTimestamp() {
+  public Long getUpdatedTimestamp() {
     return updatedTimestamp;
   }
 
-  public void setUpdatedTimestamp(Integer updatedTimestamp) {
+  public void setUpdatedTimestamp(Long updatedTimestamp) {
     this.updatedTimestamp = updatedTimestamp;
   }
